domains/category/serializer/categories.py
from rest_framework import serializers

from domains.models import Category

class CategorySerializer(serializers.ModelSerializer):

    class Meta:
        model = Category
        fields = '__all__'